Refactor Context
